+++ packages/flutter_riverpod/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased minor
+
+- Added `ChangeNotifierProvider(disposeNotifier: false)`, to disable the automatic
+  disposal of the created `ChangeNotifier`. This enables simpler migration from `pkg:provider` to `pkg:riverpod`
+  when a `ChangeNotifier` is reused between multiple providers.
+
 ## 3.2.1 - 2026-02-03
 
 - Fixed a bug where resuming a paused provider could cause it to never
@@ -1505,4 +1511,3 @@ The behavior is the same. Only the syntax changed.
 Initial release
 
 <!-- cSpell:ignoreRegExp @\w+ -->
-